rename geo_feature.py to _geo_feature.py and move types to .pyi
